requirements.txt
Flask==2.3.2
easyocr
opencv-python
numpy
google-generativeai
